pyproject.toml
[project]
name = "producer"
version = "0.1.0"
description = "Data producer for NWDAF ingestion service"
requires-python = ">=3.11"
dependencies = [
    "pandas==2.3.3",
    "fastapi==0.121.3",
    "httpx==0.28.1",
    "requests==2.32.5",
    "pyyaml==6.0.3",
    "uvicorn==0.34.0",
]

[dependency-groups]
dev = [
    "pytest==9.0.1",
]
